+++ inputguard/rules/__init__.py
@@ -1,3 +1,13 @@
 from inputguard.rules.coding import run_coding_rules
+from inputguard.rules.debug import run_debug_rules
+from inputguard.rules.optimization import run_optimization_rules
+from inputguard.rules.explanation import run_explanation_rules
+from inputguard.rules.feature import run_feature_rules
 
-__all__ = ["run_coding_rules"]
+__all__ = [
+    "run_coding_rules",
+    "run_debug_rules",
+    "run_optimization_rules",
+    "run_explanation_rules",
+    "run_feature_rules",
+]